Use span, smart pointers, stl containers to ofdm_core

+++ src/modules/ofdm/ofdm_params.h
@@ -1,14 +1,16 @@
 #pragma once
 
+#include <stdint.h>
+
 struct OFDM_Params 
 {
-    int nb_frame_symbols;
-    int nb_symbol_period;
-    int nb_null_period;
-    int nb_cyclic_prefix;
+    size_t nb_frame_symbols;
+    size_t nb_symbol_period;
+    size_t nb_null_period;
+    size_t nb_cyclic_prefix;
 
-    int nb_fft;
-    int nb_data_carriers;
+    size_t nb_fft;
+    size_t nb_data_carriers;
 
-    int freq_carrier_spacing;
+    size_t freq_carrier_spacing;
 };
